+++ uplink/builder.py
@@ -33,9 +33,6 @@ class RequestPreparer(object):
         else:
             self._session_chain = None
 
-    def _join_url_with_base(self, url):
-        return utils.urlparse.urljoin(self._base_url, url)
-
     @staticmethod
     def _get_request_hooks(contract):
         chain = list(contract.transaction_hooks)
@@ -62,7 +59,6 @@ class RequestPreparer(object):
         execution_builder.with_errbacks(self._wrap_hook(chain.handle_exception))
 
     def prepare_request(self, request_builder, execution_builder):
-        request_builder.url = self._join_url_with_base(request_builder.url)
         self._auth(request_builder)
         request_hooks = self._get_request_hooks(request_builder)
         if request_hooks: